// this.study/watchable/next/YouTube/Modern Developer Portfolio using Next js, TypeScript, Tailwind by Backbench Coder/portfolio/src/types.ts
import type { IconType } from "react-icons";

export interface IService {
  title: string;
  about: string;
  Icon: IconType;
}

export interface ISkill {
  name: string;
  level: string;
  Icon: IconType;
}

export interface IProject {
  id: number;
  name: string;
  description: string;
  imagePath: string;
  deployedUrl: string;
  githubUrl: string;
  categories: Category[];
  keyTechs: string[];
}

export type Category = "react" | "node" | "express" | "django" | "mongo";